move several sharable functions to host/src and host/include
