+++ host/CMakeLists.txt
@@ -228,3 +228,4 @@ add_crownstone_test(test_SystemTimeSync.cpp)
 add_crownstone_test(test_EventDispatcher.cpp)
 add_crownstone_test(test_BoardMap.cpp)
 add_crownstone_test(scenarios/test_ReleaseOverrideOnBehaviourUpdate.cpp)
+add_crownstone_test(scenarios/test_BehaviourConflictWithPresence.cpp)
