+++ host/test/scenarios/test_BehaviourConflictWithPresence.cpp
@@ -1,0 +1,25 @@
+/*
+ * Author: Crownstone Team
+ * Copyright: Crownstone (https://crownstone.rocks)
+ * Date: Sept 05, 2022
+ * License: LGPLv3+, Apache License 2.0, and/or MIT (triple-licensed)
+ */
+
+#include <behaviour/cs_BehaviourStore.h>
+#include <presence/cs_PresenceHandler.h>
+#include <switch/cs_SwitchAggregator.h>
+#include <events/cs_EventDispatcher.h>
+#include <presence/cs_PresenceCondition.h>
+#include <utils/date.h>
+
+int main() {
+
+    SwitchAggregator _switchAggregator;
+    BehaviourStore _behaviourStore;
+    BehaviourHandler _behaviourHandler;
+    PresenceHandler _presenceHandler;
+    SystemTime _systemTime;
+    EventDispatcher& _eventDispatcher = EventDispatcher::getInstance();
+
+    return 0;
+}
