+++ host/test/scenarios/test_ReleaseOverrideOnBehaviourUpdate.cpp
@@ -13,6 +13,9 @@
 #include <presence/cs_PresenceCondition.h>
 #include <utils/date.h>
 
+#include <testaccess/cs_SystemTime.h>
+#include <utils/cs_iostream.h>
+
 #include <cstdio>
 #include <iostream>
 #include <iomanip>
@@ -26,66 +29,6 @@
 using namespace date;
 auto now() { return std::chrono::high_resolution_clock::now(); }
 
-template<>
-class TestAccess<SystemTime> {
-public:
-	static void tick(void*) { SystemTime::tick(nullptr); }
-
-	static void fastForwardS(int seconds) {
-		 for(auto i{0}; i < seconds; i++) {
-			RTC::offsetMs(1000);
-			tick(nullptr);
-			tick(nullptr);
-			std::cout << "." << std::flush;
-		}
-	}
-
-	static void setTime(Time t) {
-		SystemTime::setTime(t.timestamp(), false, false);
-	}
-};
-
-std::ostream& operator<<(std::ostream& out, DayOfWeek t) {
-    switch (t) {
-    	case DayOfWeek::Sunday: return out << "Sunday";
-    	case DayOfWeek::Monday: return out << "Monday";
-    	case DayOfWeek::Tuesday: return out << "Tuesday";
-    	case DayOfWeek::Wednesday: return out << "Wednesday";
-    	case DayOfWeek::Thursday: return out << "Thursday";
-    	case DayOfWeek::Friday: return out << "Friday";
-    	case DayOfWeek::Saturday: return out << "Saturday";
-    }
-    return out << "UnknownDay";
-}
-
-std::ostream & operator<< (std::ostream &out, TimeOfDay t){
-	return out     << std::setw(2) << std::setfill('0') << +t.h()
-			<< ":" << std::setw(2) << std::setfill('0') << +t.m()
-			<< ":" << std::setw(2) << std::setfill('0') << +t.s();
-}
-
-std::ostream & operator<< (std::ostream &out, Time t){
-	return out << t.dayOfWeek() << " " << t.timeOfDay();
-}
-
-std::ostream & operator<< (std::ostream &out, PresenceStateDescription p){
-	int i = 0;
-	out << "rooms: {";
-	for (auto bitmask = p.getBitmask(); bitmask != 0; bitmask >>= 1) {
-		out << (i?", ":"") << i;
-		i++;
-	}
-	out << "}";
-	return out;
-}
-
-std::ostream & operator<< (std::ostream &out, std::optional<PresenceStateDescription> p_opt) {
-	if(p_opt) {
-		return out << p_opt.value();
-	} else {
-		return out << "none";
-	}
-}
 
 int main() {
 	LOGd("hello world");
